requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain_community
flask
pypdf
langchain_experimental
pinecone[grpc]
langchain-pinecone
python-dotenv
langchain_openai
langchain-huggingface
-e .